extract ResetAndWipeUseCase
